Corrección Bus y botón volver

+++ SGPP/database/migrations/2019_03_10_151029_create_asignacions_table.php
@@ -16,7 +16,7 @@ class CreateAsignacionsTable extends Migration
         Schema::create('asignacions', function (Blueprint $table) {
             $table->increments('id');
             $table->decimal('horasRealizadas');
-            $table->integer('notaFinal');
+            $table->decimal('notaFinal');
             $table->integer('alumno_id')->unsigned();
             $table->foreign('alumno_id')->references('id')->on('users');
             $table->integer('tutorAcad_id')->unsigned();
