+++ SGPP/database/migrations/2019_03_10_160013_create_evaluacions_table.php
@@ -15,7 +15,7 @@ class CreateEvaluacionsTable extends Migration
     {
         Schema::create('evaluacions', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('nota');
+            $table->decimal('nota');
             $table->string('observacion');
             $table->integer('asignacion_id')->unsigned();
             $table->foreign('asignacion_id')->references('id')->on('asignacions');
